web_app/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
click==8.3.1
fastapi==0.128.0
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
SQLAlchemy==2.0.45
starlette==0.50.0
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
asyncpg==0.31.0
jinja2==3.1.6
markupsafe==3.0.3
python-multipart==0.0.21
sqladmin==0.22.0
wtforms==3.1.2
PyJWT==2.10.1
passlib==1.7.4
redis==7.1.0
aiosmtplib==5.0.0
dnspython==2.8.0
email-validator==2.3.0